Move styles for messaging and records
