+++ src/js/messaging/sass/partials/_message-category.scss
